+++ front/src/Provider.tsx
@@ -1,0 +1,60 @@
+import { gql, useMutation } from "@apollo/client";
+import React, { useEffect, useState } from "react";
+import { ToastContainer } from "react-toastify";
+import Context, { User, UserCredentials } from "./components/context/Context";
+import "react-toastify/dist/ReactToastify.css";
+import Router from "./Router";
+
+const LOGIN = gql`
+  mutation login($input: AuthLoginInput!) {
+    login(input: $input) {
+      token
+      user {
+        email
+        nickname
+        _id
+      }
+    }
+  }
+`;
+
+const Provider = () => {
+  const [login, { data: loginData }] = useMutation(LOGIN);
+  const [user, setUser] = useState<User>(null);
+  const updateUser = (data: User) => {
+    if (data) {
+      setUser(data);
+    }
+  };
+
+  const logUser = async (userCredentials: UserCredentials) => {
+    try {
+      await login({ variables: { input: userCredentials } });
+    } catch (err: any) {
+      // eslint-disable-next-line no-console
+      console.log("err.message :>> ", err.message);
+    }
+  };
+
+  useEffect(() => {
+    if (loginData?.login.user && setUser) {
+      localStorage.setItem("jwt", JSON.stringify(loginData.login.token));
+      setUser(loginData.login.user);
+    }
+  }, [loginData, setUser]);
+
+  return (
+    <Context.Provider
+      value={{
+        user,
+        updateUser,
+        logUser,
+      }}
+    >
+      <Router />
+      <ToastContainer />
+    </Context.Provider>
+  );
+};
+
+export default Provider;
