+++ front/src/components/context/Context.tsx
@@ -16,16 +16,12 @@ const Context = React.createContext<{
   user: User;
   logUser: (userCredentials: UserCredentials) => Promise<void>;
   updateUser: (data: User) => void;
-  APICallDone: boolean;
-  updateAPICallDone: () => void;
 }>({
   user: null,
   logUser: () => {
     return Promise.resolve(undefined);
   },
   updateUser: () => {},
-  APICallDone: false,
-  updateAPICallDone: () => {},
 });
 
 export default Context;
